arena_api/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
